+++ pkg/analyzer/lib/dart/analysis/utilities.dart
@@ -23,17 +23,36 @@ import 'package:meta/meta.dart';
 ///
 /// If a [resourceProvider] is given, it will be used to access the file system.
 ///
-/// Note that if more than one file is going to be parsed then this function is
-/// inefficient. Clients should instead use [AnalysisContextCollection] to
-/// create one or more contexts and use those contexts to parse the files.
+/// [featureSet] determines what set of features will be assumed by the parser.
+/// This parameter is required because the analyzer does not yet have a
+/// performant way of computing the correct feature set for a single file to be
+/// parsed.  Callers that need the feature set to be strictly correct must
+/// create an [AnalysisContextCollection], query it to get an [AnalysisContext],
+/// query it to get an [AnalysisSession], and then call `getParsedUnit`.
 ///
-/// Deprecated: please use [parseFile2] instead.
-@Deprecated('Use parseFile2')
-ParsedUnitResult parseFile(
-    {@required String path, ResourceProvider resourceProvider}) {
-  AnalysisContext context =
-      _createAnalysisContext(path: path, resourceProvider: resourceProvider);
-  return context.currentSession.getParsedUnit(path);
+/// Callers that don't need the feature set to be strictly correct can pass in
+/// `FeatureSet.fromEnableFlags([])` to enable the default set of features; this
+/// is much more performant than using an analysis session, because it doesn't
+/// require the analyzer to process the SDK.
+///
+/// If [throwIfDiagnostics] is `true` (the default), then if any diagnostics are
+/// produced because of syntactic errors in the [content] an `ArgumentError`
+/// will be thrown. If the parameter is `false`, then the caller can check the
+/// result to see whether there are any errors.
+ParseStringResult parseFile(
+    {@required String path,
+    ResourceProvider resourceProvider,
+    @required FeatureSet featureSet,
+    bool throwIfDiagnostics: true}) {
+  if (featureSet == null) {
+    throw ArgumentError('A non-null feature set must be provided.');
+  }
+  resourceProvider ??= PhysicalResourceProvider.INSTANCE;
+  var content = (resourceProvider.getResource(path) as File).readAsStringSync();
+  return parseString(
+      content: content,
+      featureSet: featureSet,
+      throwIfDiagnostics: throwIfDiagnostics);
 }
 
 /// Return the result of parsing the file at the given [path].
@@ -56,18 +75,15 @@ ParsedUnitResult parseFile(
 /// produced because of syntactic errors in the [content] an `ArgumentError`
 /// will be thrown. If the parameter is `false`, then the caller can check the
 /// result to see whether there are any errors.
+@Deprecated('Use parseFile')
 ParseStringResult parseFile2(
     {@required String path,
     ResourceProvider resourceProvider,
     @required FeatureSet featureSet,
     bool throwIfDiagnostics: true}) {
-  if (featureSet == null) {
-    throw ArgumentError('A non-null feature set must be provided.');
-  }
-  resourceProvider ??= PhysicalResourceProvider.INSTANCE;
-  var content = (resourceProvider.getResource(path) as File).readAsStringSync();
-  return parseString(
-      content: content,
+  return parseFile(
+      path: path,
+      resourceProvider: resourceProvider,
       featureSet: featureSet,
       throwIfDiagnostics: throwIfDiagnostics);
 }