+++ pkg/analyzer/lib/dart/ast/ast_factory.dart
@@ -151,19 +151,21 @@ abstract class AstFactory {
   /// can be `null` if the reference is not to a constructor.
   CommentReference commentReference(Token newKeyword, Identifier identifier);
 
-  /// Returns a newly created compilation unit to have the given directives
-  /// and declarations. The [scriptTag] can be `null` if there is no script tag
-  /// in the compilation unit. The list of [directives] can be `null` if there
-  /// are no directives in the compilation unit. The list of [declarations] can
-  /// be `null` if there are no declarations in the compilation unit.
-  @Deprecated('Use compilationUnit2')
+  /// Returns a newly created compilation unit to have the given directives and
+  /// declarations.  The [scriptTag] can be `null` (or omitted) if there is no
+  /// script tag in the compilation unit.  The list of [declarations] can be
+  /// `null` (or omitted) if there are no directives in the compilation unit.
+  /// The list of `declarations` can be `null` (or omitted) if there are no
+  /// declarations in the compilation unit.  The [featureSet] can be `null` if
+  /// the set of features for this compilation unit is not known (this
+  /// restricts what analysis can be done of the compilation unit).
   CompilationUnit compilationUnit(
-      Token beginToken,
+      {@required Token beginToken,
       ScriptTag scriptTag,
       List<Directive> directives,
       List<CompilationUnitMember> declarations,
-      Token endToken,
-      [FeatureSet featureSet]);
+      @required Token endToken,
+      @required FeatureSet featureSet});
 
   /// Returns a newly created compilation unit to have the given directives and
   /// declarations.  The [scriptTag] can be `null` (or omitted) if there is no
@@ -173,6 +175,7 @@ abstract class AstFactory {
   /// declarations in the compilation unit.  The [featureSet] can be `null` if
   /// the set of features for this compilation unit is not known (this
   /// restricts what analysis can be done of the compilation unit).
+  @Deprecated('Use compilationUnit')
   CompilationUnit compilationUnit2(
       {@required Token beginToken,
       ScriptTag scriptTag,
@@ -463,16 +466,6 @@ abstract class AstFactory {
 
   /// Returns a newly created for statement.
   ForStatement forStatement(
-      {Token awaitKeyword,
-      Token forKeyword,
-      Token leftParenthesis,
-      ForLoopParts forLoopParts,
-      Token rightParenthesis,
-      Statement body});
-
-  /// Returns a newly created for statement.
-  @Deprecated('Replaced by forStatement')
-  ForStatement forStatement2(
       {Token awaitKeyword,
       Token forKeyword,
       Token leftParenthesis,